Move ReadOnlyDictionary from System.ObjectModel to System.Private.CoreLib
