+++ src/libraries/System.ObjectModel/src/System.ObjectModel.csproj
@@ -12,7 +12,6 @@
     <Compile Include="System\Collections\Specialized\NotifyCollectionChangedEventArgs.cs" />
     <Compile Include="System\Collections\ObjectModel\KeyedCollection.cs" />
     <Compile Include="System\Collections\ObjectModel\ObservableCollection.cs" />
-    <Compile Include="System\Collections\ObjectModel\ReadOnlyDictionary.cs" />
     <Compile Include="System\Collections\ObjectModel\ReadOnlyObservableCollection.cs" />
     <Compile Include="System\ComponentModel\DataErrorsChangedEventArgs.cs" />
     <Compile Include="System\ComponentModel\INotifyDataErrorInfo.cs" />
