+++ src/libraries/System.ObjectModel/src/System/Collections/CollectionHelpers.cs
@@ -34,40 +34,5 @@ namespace System.Collections
                 throw new ArgumentException(SR.Arg_ArrayPlusOffTooSmall);
             }
         }
-
-        internal static void CopyTo<T>(ICollection<T> collection, Array array, int index)
-        {
-            ValidateCopyToArguments(collection.Count, array, index);
-
-            if (collection is ICollection nonGenericCollection)
-            {
-                // Easy out if the ICollection<T> implements the non-generic ICollection
-                nonGenericCollection.CopyTo(array, index);
-            }
-            else if (array is T[] items)
-            {
-                collection.CopyTo(items, index);
-            }
-            else
-            {
-                // We can't cast array of value type to object[], so we don't support widening of primitive types here.
-                if (array is not object?[] objects)
-                {
-                    throw new ArgumentException(SR.Argument_InvalidArrayType, nameof(array));
-                }
-
-                try
-                {
-                    foreach (T item in collection)
-                    {
-                        objects[index++] = item;
-                    }
-                }
-                catch (ArrayTypeMismatchException)
-                {
-                    throw new ArgumentException(SR.Argument_InvalidArrayType, nameof(array));
-                }
-            }
-        }
     }
 }
