+++ src/libraries/System.ObjectModel/src/System/Collections/Generic/DebugView.cs
@@ -25,25 +25,4 @@ namespace System.Collections.Generic
             }
         }
     }
-
-    internal sealed class DictionaryDebugView<K, V> where K: notnull
-    {
-        private readonly IDictionary<K, V> _dict;
-
-        public DictionaryDebugView(IDictionary<K, V> dictionary)
-        {
-            _dict = dictionary ?? throw new ArgumentNullException(nameof(dictionary));
-        }
-
-        [DebuggerBrowsable(DebuggerBrowsableState.RootHidden)]
-        public KeyValuePair<K, V>[] Items
-        {
-            get
-            {
-                KeyValuePair<K, V>[] items = new KeyValuePair<K, V>[_dict.Count];
-                _dict.CopyTo(items, 0);
-                return items;
-            }
-        }
-    }
 }
